Métodos basicos completados, cambios con mapeadores y usos de las DTOs. Añadidas queries a los repositorios, lógica de la creación de chats, notificaciones, notas, observaciones, material escolar y equipos. Conexión con el frontend y la BD. :sparkles:

+++ src/main/java/com/spring/learnsphere/dto/AsignaturaDTO.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Data
 public class AsignaturaDTO {
-    private Long asignaturaId;
+    private Integer asignaturaId;
     private String nombre;
 }
